feat(rengine/codegen): refactor and add a new kind of visitor
